add options for video/audio only playback
The -v and -a options can be used to start a video or audio
only playback.  The -r option can be used to set video
playback rate.  This is helpful when playing files whose
video frames are not followed by exactly one audio frame;
the video and audio playback can be done in separate
processes.