ffs: fix the uninitialized avpacket in ffs_adec()
Ffmpeg aac decoder codec started using some of the fields of the passed
afpacket since some version of ffmpeg.  This resulted in fbff segfaults
since it initialized only data and size fields of afpacket.  This patch
initializes all fields of the passed packet.

Because of this bug, fbff could not play files with aac audio streams in
the last few months.  Since other audio formats were decoded perfectly, I
came to the conclusion that there is something wrong with the aac decoder
of ffmpeg.  This delayed fixing the bug till now.  I hoped ffmpeg had
a saner API.

Thanks to Sara for helping me find this annoying bug.

Reported-by: Sara Fauzia <sfauzia@fas.harvard.edu>